refactor: HeaderFallback 위치 Header 하위로 이동
